+++ app/scripts/drsPlatform.service.js
@@ -4,11 +4,20 @@ angular
   .module('drsApp')
   .factory('PlatformService', PlatformService);
 
-function PlatformService() {
+PlatformService.$inject = ['$rootScope', 'SettingsService'];
+
+function PlatformService($rootScope, SettingsService) {
   const platforms = ['PS4', 'XB1'];
   const service = {
     active: platforms[0],
     platforms: platforms
   };
+
+  $rootScope.$on("drs-settings-loaded", function() {
+    service.active = (platforms.indexOf(SettingsService.platform) !== -1)
+      ? SettingsService.platform
+      : platforms[0];
+  });
+
   return service;
 }